메세지 창 구현, 몬스터 피격 애니메이션 추가

+++ Client/GameCoding/UI.cpp
@@ -41,10 +41,10 @@ RECT UI::GetRect()
 {
 	RECT rect =
 	{
-		_pos.x - _size.x / 2,
-		_pos.y - _size.y / 2,
-		_pos.x + _size.x / 2,
-		_pos.y + _size.y / 2,
+		(int)_pos.x - (int)_size.x / 2,
+		(int)_pos.y - (int)_size.y / 2,
+		(int)_pos.x + (int)_size.x / 2,
+		(int)_pos.y + (int)_size.y / 2,
 	};
 
 	return rect;
